refactor(frontend): 401 认证错误自动跳转登录
- 新增 ApiClient 类统一处理 API 请求
- 实现自动处理 401 认证错误并跳转登录
- 优化请求配置和错误处理
- 将 API 请求方法抽象为通用方法
- 新增文件上传通用方法
